Ensure we have a kind for each message.
Ensure that each message has a kind by requiring that it is supplied on
the command line as a parameter or that heuristic kind guessing is
turned on (through the --guess flag). The kinds are now implemented as
an Enum. kind_is_func() is now a pure predicate which doesn't add new
attributes.