Encode message data to UTF-8 bytes (if not).
The message data should really be a raw bytestream since we support
stuff other than "text" (Unicode/ASCII bytestreams), but since we
currently read data as a command-line parameter (and possibly want to
retain this feature as an option in the future), it gets automatically
decoded to str. This is a problem only for kind blob (since URLs are
always going to be readable bytestreams, hopefully) so we need to check
for this case and convert back to bytes if needed.