Use Content-Type instead of downloading implicitly
Not a good idea to download files implicitly so we shouldn't use magic
as a default fallback way of determining the mimetype. Use the HEAD HTTP
method instead and read the Content-Type header.

Explicitly downloading the file and using magic on it should still be
possible, though, but is unsupported at the moment since we're not sure
how we'll provide that feature.