core: Use ElasticDict for messages and merge them with match_group.
Messages are now implemented with ElasticDict. This enables us to track
regex match groups created by some clause inside the message itself
therefore obviating the need for passing the match_group variable
separately. The message is then reset prior to advancing to the next
rule so it receives the clean, original version of the message.

This paves the way for implementing passing of the message from a parent
rule to a different, child rule (i.e. "replumb") with the parent rule
setting some variables on the message for the child's consumption. If
the child rule doesn't match the message, we will still be able to reset
the message and continue matching with the next rule with a clean
version of the message.

Also, istype was changed to be able to set match groups too for symmetry
since it already does regex matching anyway.