parser: Avoid using <parser>.parseFile.
If parseFile is used to parse files, an annoying bug occurs if the error
is in the (second to) last line of the file and it ends with a newline
(as it should). pyparsing then reports the error in the subsequent empty
line. The workaround is to read in the file manually, rstrip its
whitespace and then pass it to parseString instead.

argparse doesn't open the rules file automatically anymore so we can
give uniform error messages for missing rules files.